# backend/openapi/global-openapi.yaml
openapi: 3.1.0
info:
  title: Single Window API
  version: 1.0.0
paths:
...
  /api/v1/revisions:
    get:
      tags:
        - Revisions
      summary: Получить список ревизий
      operationId: GetRevisionsV1
      parameters:
        - in: query
          name: disputeId
          schema:
            type: string
          required: false
          description: ID Спора для которого необходимо получить ревизии (если указан и disputeId, и status, то выборка осуществляется по disputeId)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: Смещение с которого необходимо получить ревизии
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10
          required: false
          description: Количество возращаемых ревизий
        - in: query
          name: status
          required: false
          schema:
            $ref: '#/components/schemas/EntityStatus'
          description: Статус возвращаемых ревизий (если указан и disputeId, и status, то выборка осуществляется по disputeId)
      responses:
        '200':
          description: Возвращает массив объектов "Revision"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Revision'
      security:
        - SWCookieAuth: []
  /api/v1/revision:
    get:
      tags:
        - Revisions
      summary: Получить ревизию по revisionId
      operationId: GetRevisionV1
      parameters:
        - in: query
          name: revisionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Вовращает данные по ревизии
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/Revision'
      security:
        - SWCookieAuth: []
    post:
      tags:
        - Revisions
      summary: Создание новой ревизии
      operationId: CreateRevisionV1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - dispute_id
                - organization_id
                - message_body
              properties:
                dispute_id:
                  type: string
                organization_id:
                  type: string
                message_body:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Возращает данные по созданной ревизии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
      security:
        - SWCookieAuth: []
...
  /api/v1/revision/correspondences:
    get:
      tags:
        - Revisions
      summary: Получить список кореспонденций по revision id
      operationId: GetCorrespondencesV1
      parameters:
        - in: query
          name: revision_id
          required: true
          schema:
            type: string
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10
          required: false
      responses:
        '200':
          description: Возвращает список кореспонденций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Correspondence'
      security:
        - SWCookieAuth: []
components:
  schemas:
    Correspondence:
      type: object
      required:
        - correspondence_id
        - revision_id
        - sender_id
        - created_at
      properties:
        correspondence_id:
          type: string
        revision_id:
          type: string
        sender_id:
          type: string
        created_at:
          type: string
          format: date-time
        message_body:
          type: string
        attachment_path:
          type: string
...
    EntityStatus:
      type: string
      enum:
        - opened
        - in_work
        - closed
        - any
      default: any
...
    Revision:
      type: object
      required:
        - revision_id
        - organization_id
        - dispute_id
        - status
        - created_at
      properties:
        revision_id:
          type: string
        organization_title:
          type: string
        organization_id:
          type: string
        dispute_id:
          type: string
        status:
          $ref: '#/components/schemas/EntityStatus'
        created_at:
          type: string
          format: date-time
        in_work_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        worker_name:
          type: string
        worker_id:
          type: string
...
  securitySchemes:
    SWCookieAuth:
      type: apiKey
      in: cookie
      name: swCookieAuth